requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
whitenoise==6.8.2
